add some unit tests for debezium_avro
