+++ src/connector/src/test_data/debezium_avro_msg_schema.avsc
@@ -1,0 +1,190 @@
+{
+    "type": "record",
+    "name": "Envelope",
+    "namespace": "dbserver1.inventory.customers",
+    "fields": [
+        {
+            "name": "before",
+            "type": [
+                "null",
+                {
+                    "type": "record",
+                    "name": "Value",
+                    "fields": [
+                        {
+                            "name": "id",
+                            "type": "int"
+                        },
+                        {
+                            "name": "first_name",
+                            "type": "string"
+                        },
+                        {
+                            "name": "last_name",
+                            "type": "string"
+                        },
+                        {
+                            "name": "email",
+                            "type": "string"
+                        }
+                    ],
+                    "connect.name": "dbserver1.inventory.customers.Value"
+                }
+            ],
+            "default": null
+        },
+        {
+            "name": "after",
+            "type": [
+                "null",
+                "Value"
+            ],
+            "default": null
+        },
+        {
+            "name": "source",
+            "type": {
+                "type": "record",
+                "name": "Source",
+                "namespace": "io.debezium.connector.mysql",
+                "fields": [
+                    {
+                        "name": "version",
+                        "type": "string"
+                    },
+                    {
+                        "name": "connector",
+                        "type": "string"
+                    },
+                    {
+                        "name": "name",
+                        "type": "string"
+                    },
+                    {
+                        "name": "ts_ms",
+                        "type": "long"
+                    },
+                    {
+                        "name": "snapshot",
+                        "type": [
+                            {
+                                "type": "string",
+                                "connect.version": 1,
+                                "connect.parameters": {
+                                    "allowed": "true,last,false,incremental"
+                                },
+                                "connect.default": "false",
+                                "connect.name": "io.debezium.data.Enum"
+                            },
+                            "null"
+                        ],
+                        "default": "false"
+                    },
+                    {
+                        "name": "db",
+                        "type": "string"
+                    },
+                    {
+                        "name": "sequence",
+                        "type": [
+                            "null",
+                            "string"
+                        ],
+                        "default": null
+                    },
+                    {
+                        "name": "table",
+                        "type": [
+                            "null",
+                            "string"
+                        ],
+                        "default": null
+                    },
+                    {
+                        "name": "server_id",
+                        "type": "long"
+                    },
+                    {
+                        "name": "gtid",
+                        "type": [
+                            "null",
+                            "string"
+                        ],
+                        "default": null
+                    },
+                    {
+                        "name": "file",
+                        "type": "string"
+                    },
+                    {
+                        "name": "pos",
+                        "type": "long"
+                    },
+                    {
+                        "name": "row",
+                        "type": "int"
+                    },
+                    {
+                        "name": "thread",
+                        "type": [
+                            "null",
+                            "long"
+                        ],
+                        "default": null
+                    },
+                    {
+                        "name": "query",
+                        "type": [
+                            "null",
+                            "string"
+                        ],
+                        "default": null
+                    }
+                ],
+                "connect.name": "io.debezium.connector.mysql.Source"
+            }
+        },
+        {
+            "name": "op",
+            "type": "string"
+        },
+        {
+            "name": "ts_ms",
+            "type": [
+                "null",
+                "long"
+            ],
+            "default": null
+        },
+        {
+            "name": "transaction",
+            "type": [
+                "null",
+                {
+                    "type": "record",
+                    "name": "block",
+                    "namespace": "event",
+                    "fields": [
+                        {
+                            "name": "id",
+                            "type": "string"
+                        },
+                        {
+                            "name": "total_order",
+                            "type": "long"
+                        },
+                        {
+                            "name": "data_collection_order",
+                            "type": "long"
+                        }
+                    ],
+                    "connect.version": 1,
+                    "connect.name": "event.block"
+                }
+            ],
+            "default": null
+        }
+    ],
+    "connect.version": 1,
+    "connect.name": "dbserver1.inventory.customers.Envelope"
+}
